feat: extract Storefront Next extension and add extension pack (Phase 2)
Moves Page Designer Assistant and related Storefront Next code out of
the core extension into a new b2c-dx-sfnext extension, and adds a
b2c-dx extension pack that installs both.

- New package: packages/vscode/storefront-next (b2c-dx-sfnext)
- New package: packages/vscode/extension-pack (b2c-dx meta-package)
- Core extension slimmed: Page Designer command moved to sfnext
- Updates publish workflow, launch/tasks config, and changeset scopes
