+++ .github/workflows/publish.yml
@@ -163,16 +163,25 @@ jobs:
           check_package "@salesforce/b2c-dx-mcp" "packages/b2c-dx-mcp" "mcp"
           check_package "@salesforce/mrt-utilities" "packages/mrt-utilities" "mrt"
 
-          # VS Code extension — compare against git tags (not npm)
-          LOCAL_VSX_VERSION=$(node -p "require('./packages/vscode/core/package.json').version")
-          LAST_VSX_TAG=$(git tag -l "b2c-dx-core@*" --sort=-v:refname | grep -v '@latest$' | head -1 | sed 's/b2c-dx-core@//')
-          echo "b2c-dx-core: local=${LOCAL_VSX_VERSION} tag=${LAST_VSX_TAG:-none}"
-          if [ "$LOCAL_VSX_VERSION" != "$LAST_VSX_TAG" ]; then
-            echo "publish_vsx=true" >> $GITHUB_OUTPUT
-            echo "version_vsx=${LOCAL_VSX_VERSION}" >> $GITHUB_OUTPUT
-          else
-            echo "publish_vsx=false" >> $GITHUB_OUTPUT
-          fi
+          # VS Code extensions — compare against git tags (not npm)
+          check_vsx_package() {
+            local pkg_name=$1
+            local pkg_path=$2
+            local output_key=$3
+
+            LOCAL_VERSION=$(node -p "require('./${pkg_path}/package.json').version")
+            LAST_TAG=$(git tag -l "${pkg_name}@*" --sort=-v:refname | grep -v '@latest$' | head -1 | sed "s/${pkg_name}@//")
+            echo "${pkg_name}: local=${LOCAL_VERSION} tag=${LAST_TAG:-none}"
+            if [ "$LOCAL_VERSION" != "$LAST_TAG" ]; then
+              echo "publish_${output_key}=true" >> $GITHUB_OUTPUT
+              echo "version_${output_key}=${LOCAL_VERSION}" >> $GITHUB_OUTPUT
+            else
+              echo "publish_${output_key}=false" >> $GITHUB_OUTPUT
+            fi
+          }
+
+          check_vsx_package "b2c-dx-core" "packages/vscode/core" "vsx"
+          check_vsx_package "b2c-dx-sfnext" "packages/vscode/storefront-next" "vsx_sfnext"
 
           # Check if docs version changed (private package — not published to npm, uses git tag)
           DOCS_VERSION=$(node -p "require('./docs/package.json').version")
@@ -205,7 +214,7 @@ jobs:
 
       - name: Run tests
         if: steps.release-type.outputs.type == 'nightly' || (steps.changesets.outputs.skip != 'true' && steps.quick-check.outputs.skip != 'true')
-        run: pnpm --filter '!b2c-dx-core' run test
+        run: pnpm --filter '!b2c-dx-core' --filter '!b2c-dx-sfnext' --filter '!b2c-dx' run test
 
       - name: Publish SDK to npm
         if: steps.release-type.outputs.type == 'nightly' || steps.packages.outputs.publish_sdk == 'true'
@@ -231,10 +240,9 @@ jobs:
           pnpm --filter @salesforce/mrt-utilities publish --provenance --no-git-checks
           --tag ${{ steps.release-type.outputs.type == 'nightly' && 'nightly' || steps.packages.outputs.tag_mrt }}
 
-      - name: Package VS Code extension
-        if: steps.release-type.outputs.type == 'stable' && steps.packages.outputs.publish_vsx == 'true'
-        working-directory: packages/vscode/core
-        run: pnpm run package
+      - name: Package VS Code extensions
+        if: steps.release-type.outputs.type == 'stable' && (steps.packages.outputs.publish_vsx == 'true' || steps.packages.outputs.publish_vsx_sfnext == 'true')
+        run: pnpm run package:extensions
 
       - name: Create git tags
         if: steps.release-type.outputs.type == 'stable' && steps.changesets.outputs.skip != 'true' && steps.quick-check.outputs.skip != 'true'
@@ -270,6 +278,12 @@ jobs:
 
           if [[ "${{ steps.packages.outputs.publish_vsx }}" == "true" ]]; then
             TAG="b2c-dx-core@${{ steps.packages.outputs.version_vsx }}"
+            git tag "$TAG"
+            TAGS_CREATED="$TAGS_CREATED $TAG"
+          fi
+
+          if [[ "${{ steps.packages.outputs.publish_vsx_sfnext }}" == "true" ]]; then
+            TAG="b2c-dx-sfnext@${{ steps.packages.outputs.version_vsx_sfnext }}"
             git tag "$TAG"
             TAGS_CREATED="$TAGS_CREATED $TAG"
           fi
@@ -438,7 +452,7 @@ jobs:
             exit 0
           fi
 
-          gh release upload "$RELEASE_TAG" packages/vscode/core/*.vsix
+          gh release upload "$RELEASE_TAG" packages/vscode/core/*.vsix packages/vscode/storefront-next/*.vsix || true
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
