+++ .vscode/launch.json
@@ -6,11 +6,14 @@
       "type": "extensionHost",
       "request": "launch",
       "args": [
-        "--extensionDevelopmentPath=${workspaceFolder}/packages/vscode/core"
+        "--extensionDevelopmentPath=${workspaceFolder}/packages/vscode/core",
+        "--extensionDevelopmentPath=${workspaceFolder}/packages/vscode/storefront-next"
       ],
       "outFiles": [
-        "${workspaceFolder}/packages/vscode/core/dist/**/*.js"
-      ]
+        "${workspaceFolder}/packages/vscode/core/dist/**/*.js",
+        "${workspaceFolder}/packages/vscode/storefront-next/dist/**/*.js"
+      ],
+      "preLaunchTask": "build extensions"
     }
   ]
 }
