+++ packages/vscode/core/src/extension.ts
@@ -6,7 +6,6 @@
 import {DwJsonSource} from '@salesforce/b2c-tooling-sdk/config';
 import {configureLogger, getLogger} from '@salesforce/b2c-tooling-sdk/logging';
 
-import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import type {B2CDXApi} from './api.js';
@@ -19,69 +18,6 @@ import {registerSandboxTree} from './sandbox-tree/index.js';
 import {registerScaffold} from './scaffold/index.js';
 import {registerApiBrowser} from './api-browser/index.js';
 import {registerWebDavTree} from './webdav-tree/index.js';
-
-function getWebviewContent(context: vscode.ExtensionContext): string {
-  const htmlPath = path.join(context.extensionPath, 'src', 'webview.html');
-  return fs.readFileSync(htmlPath, 'utf-8');
-}
-
-/** PascalCase for use in template content (class names, types, etc.). e.g. "first page" → "FirstPage" */
-function pageNameToPageId(pageName: string): string {
-  return pageName
-    .trim()
-    .split(/\s+/)
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-    .join('');
-}
-
-/** camelCase for filename. e.g. "first page" → "firstPage" */
-function pageNameToFileNameId(pageName: string): string {
-  const pascal = pageNameToPageId(pageName || 'Page');
-  return pascal.charAt(0).toLowerCase() + pascal.slice(1);
-}
-
-type RegionForm = {id: string; name: string; description: string; maxComponents: number};
-
-type WebviewMessage =
-  | {type: 'openExternal'}
-  | {
-      type: 'submitForm';
-      pageType: {name?: string; description?: string; supportedAspectTypes?: string[]};
-      regions: RegionForm[];
-    };
-
-function renderTemplate(
-  template: string,
-  pageName: string,
-  pageDescription: string,
-  supportedAspectTypes: string[],
-  regions: RegionForm[],
-): string {
-  const pageId = pageNameToPageId(pageName || 'Page');
-  const quoted = (s: string) => `'${String(s).replace(/'/g, "\\'")}'`;
-  const aspectsStr = `[${supportedAspectTypes.map((a) => quoted(a)).join(', ')}]`;
-  const regionsBlock = regions
-    .map(
-      (r) =>
-        `{
-        id: ${quoted(r.id)},
-        name: ${quoted(r.name)},
-        description: ${quoted(r.description)},
-        maxComponents: ${r.maxComponents},
-    }`,
-    )
-    .join(',\n    ');
-  const firstRegionId = regions[0]?.id ?? '';
-
-  return template
-    .replace(/\$\{pageName\}/g, quoted(pageName || ''))
-    .replace(/\$\{pageDescription\}/g, quoted(pageDescription || ''))
-    .replace(/\$\{supportedAspectTypes\}/g, aspectsStr)
-    .replace('__REGIONS__', regionsBlock)
-    .replace(/\$\{pageId\}/g, pageId)
-    .replace(/\$\{pageName\}Data/g, `${pageId}Data`)
-    .replace(/\$\{regions\[0\]\.id\}/g, firstRegionId);
-}
 
 function applyLogLevel(log: vscode.OutputChannel): void {
   const config = vscode.workspace.getConfiguration('b2c-dx');
@@ -124,11 +60,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<B2CDXA
       log.show();
       vscode.window.showErrorMessage(`B2C DX activation error: ${message}`);
     };
-    context.subscriptions.push(
-      vscode.commands.registerCommand('b2c-dx.openUI', showActivationError),
-      vscode.commands.registerCommand('b2c-dx.promptAgent', showActivationError),
-      vscode.commands.registerCommand('b2c-dx.listWebDav', showActivationError),
-    );
+    context.subscriptions.push(vscode.commands.registerCommand('b2c-dx.listWebDav', showActivationError));
     return undefined;
   }
 }
@@ -142,98 +74,6 @@ async function activateInner(context: vscode.ExtensionContext, log: vscode.Outpu
   const configProvider = new B2CExtensionConfig(log, context.workspaceState);
   context.subscriptions.push(configProvider);
   await configProvider.ensureResolved();
-
-  const disposable = vscode.commands.registerCommand('b2c-dx.openUI', () => {
-    vscode.window.showInformationMessage('B2C DX: Opening Page Designer Assistant.');
-
-    const panel = vscode.window.createWebviewPanel(
-      'b2c-dx-page-designer-ui',
-      'My Extension UI',
-      vscode.ViewColumn.One,
-      {enableScripts: true},
-    );
-
-    panel.webview.html = getWebviewContent(context);
-
-    panel.webview.onDidReceiveMessage(async (msg: WebviewMessage) => {
-      if (msg.type === 'openExternal') {
-        await vscode.env.openExternal(vscode.Uri.parse('https://example.com'));
-      }
-      if (msg.type === 'submitForm') {
-        try {
-          const {pageType, regions} = msg;
-          const pageName = pageType?.name ?? '';
-          const templatePath = path.join(context.extensionPath, 'src', 'template', '_app.pageId.tsx');
-          const template = fs.readFileSync(templatePath, 'utf-8');
-          const content = renderTemplate(
-            template,
-            pageName,
-            pageType?.description ?? '',
-            pageType?.supportedAspectTypes ?? [],
-            regions ?? [],
-          );
-
-          const fileNameId = pageNameToFileNameId(pageName);
-          const fileName = `_app.${fileNameId}.tsx`;
-
-          let targetUri: vscode.Uri;
-          if (vscode.workspace.workspaceFolders?.length) {
-            const rootUri = vscode.Uri.file(configProvider.getWorkingDirectory());
-            const routesUri = vscode.Uri.joinPath(rootUri, 'routes');
-            const routesPath = routesUri.fsPath;
-            const hasRoutesFolder = fs.existsSync(routesPath) && fs.statSync(routesPath).isDirectory();
-            targetUri = hasRoutesFolder
-              ? vscode.Uri.joinPath(routesUri, fileName)
-              : vscode.Uri.joinPath(rootUri, fileName);
-          } else {
-            const picked = await vscode.window.showSaveDialog({
-              defaultUri: vscode.Uri.joinPath(context.globalStorageUri, fileName),
-              saveLabel: 'Create file',
-            });
-            if (!picked) {
-              return;
-            }
-            targetUri = picked;
-          }
-
-          vscode.window.showInformationMessage(`Writing file to: ${targetUri.fsPath}`);
-
-          await vscode.workspace.fs.writeFile(targetUri, Buffer.from(content, 'utf-8'));
-          await vscode.window.showInformationMessage(`Saved to: ${targetUri.fsPath}`, 'Open');
-          const doc = await vscode.workspace.openTextDocument(targetUri);
-          await vscode.window.showTextDocument(doc, {
-            viewColumn: panel.viewColumn ?? vscode.ViewColumn.One,
-            preview: false,
-            preserveFocus: false,
-          });
-        } catch (err) {
-          const message = err instanceof Error ? err.message : String(err);
-          vscode.window.showErrorMessage(`Failed to save: ${message}`);
-        }
-      }
-    });
-  });
-
-  const promptAgentDisposable = vscode.commands.registerCommand('b2c-dx.promptAgent', async () => {
-    const prompt = await vscode.window.showInputBox({
-      title: 'Prompt Agent',
-      placeHolder: 'Enter your prompt for the agent...',
-    });
-    if (prompt === undefined || prompt === '') {
-      return;
-    }
-    try {
-      await vscode.env.clipboard.writeText(prompt);
-      await vscode.commands.executeCommand('composer.newAgentChat');
-      await new Promise((resolve) => setTimeout(resolve, 300));
-      await vscode.commands.executeCommand('editor.action.clipboardPasteAction');
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      vscode.window.showWarningMessage(
-        `Could not open Cursor chat: ${message}. Run this extension in Cursor to send prompts to the agent.`,
-      );
-    }
-  });
 
   const listWebDavDisposable = vscode.commands.registerCommand('b2c-dx.listWebDav', () => {
     vscode.commands.executeCommand('b2cWebdavExplorer.focus');
@@ -407,8 +247,6 @@ async function activateInner(context: vscode.ExtensionContext, log: vscode.Outpu
   });
 
   context.subscriptions.push(
-    disposable,
-    promptAgentDisposable,
     listWebDavDisposable,
     instanceStatusBar,
     instanceConfigRegistration,
