+++ packages/vscode/storefront-next/.gitignore
@@ -1,0 +1,6 @@
+out
+dist
+node_modules
+.vscode-test/
+.vsce-stage/
+*.vsix
