+++ packages/vscode/storefront-next/.vscodeignore
@@ -1,9 +1,11 @@
 .vscode/**
 .vscode-test/**
 src/**
-!src/api-browser/swagger-webview.html
+!src/page-designer/webview.html
+!src/page-designer/template/**
 out/**
 **/node_modules/**
+
 .gitignore
 **/tsconfig.json
 **/eslint.config.mjs
