+++ packages/vscode/storefront-next/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"exclude": ["src/page-designer/template", "src/test"],
+	"compilerOptions": {
+		"module": "Node16",
+		"target": "ES2022",
+		"outDir": "out",
+		"lib": ["ES2022"],
+		"sourceMap": true,
+		"rootDir": "src",
+		"strict": true,
+		"skipLibCheck": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true
+	}
+}
